isolate group notification logic in new files
